Updated unix driver to load appropriate driver runtime, addressed some other comments
