+++ src/System.Device.Gpio/Interop/Unix/SafeChipIteratorHandle.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Runtime.InteropServices;
+using System.Text;
+
+namespace System.Device.Gpio
+{
+    class SafeChipIteratorHandle : SafeHandle
+    {
+        internal SafeChipIteratorHandle() : base(IntPtr.Zero, true) { }
+
+        protected override bool ReleaseHandle()
+        {
+            Interop.gpiod_chip_iter_free(handle);
+            return true;
+        }
+
+        public override bool IsInvalid => handle == IntPtr.Zero || handle == new IntPtr(-1);
+    }
+}
