--- src/System.Device.Gpio/Interop/Unix/libgpiod/Interop.gpiod_chip_close.cs
@@ -1,19 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information
-
-using System;
-using System.Runtime.InteropServices;
-
-internal partial class Interop
-{
-    private const string LibgpiodLibrary = "libgpiod";
-
-    /// <summary>
-    ///     Close a GPIO chip handle and release all allocated resources.
-    ///     <see href="https://git.kernel.org/pub/scm/libs/libgpiod/libgpiod.git/tree/include/gpiod.h#n443">here</see>.
-    /// </summary>
-    /// <param name="chip">The GPIO chip pointer</param>
-    [DllImport(LibgpiodLibrary)]
-    internal static extern void gpiod_chip_close(IntPtr chip);
-}
