+++ src/System.Device.Gpio/System/Device/Gpio/GpioController.Linux.cs
@@ -40,6 +40,7 @@ namespace System.Device.Gpio
                     {
                         if (match.Groups[1].Value == RaspberryPiHardware)
                         {
+                            Console.WriteLine("raspberry driver");
                             return new RaspberryPi3Driver();
                         }
                         // Commenting out as HummingBoard driver is not implemented yet, will be added back after implementation 
@@ -48,11 +49,11 @@ namespace System.Device.Gpio
                         //{
                         //    return new HummingBoardDriver();
                         //} 
-                        return new UnixDriver();
+                        return UnixDriver.InitUnixDriver();
                     }
                 }
             }
-            return new UnixDriver();
+            return UnixDriver.InitUnixDriver();
         }
     }
 }
